backend/requirements.txt
# Backend dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
firebase-admin>=6.4.0
Pillow>=10.2.0
numpy>=1.26.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
